* better-deploy:
  Make deploy.sh be able to run for non-production environments
  Remove hardcoded env in deploy.sh

+++ bin/deploy.sh
@@ -7,23 +7,29 @@ set -ex
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 cd "$DIR"/..
 
-# Update repository
-git checkout master
-git fetch -tp
-git pull
+source .env
+
+if [  "$ENV" = "production" ]; then
+    # Update repository
+    git checkout master
+    git fetch -tp
+    git pull
+fi
 
 # Build and start container
-docker build -t $PROJECT_NAME:production .
+docker build -t $PROJECT_NAME:$ENV .
 docker stop $PROJECT_NAME || echo
 docker container prune -f
 docker run \
     --detach \
     --restart=always \
     --publish=127.0.0.1:$INTERNAL_PORT:$INTERNAL_PORT \
-    --name=$PROJECT_NAME $PROJECT_NAME:production
+    --name=$PROJECT_NAME $PROJECT_NAME:$ENV
 
-# Cleanup docker
-docker image prune -f --filter "until=336h"
+if [  "$ENV" = "production" ]; then
+    # Cleanup docker
+    docker image prune -f --filter "until=336h"
 
-# Update nginx
-sudo service nginx reload
+    # Update nginx
+    sudo service nginx reload
+fi
